refactor(merge): consolidate duplicated helpers across the merge family
The five stax merge command modules re-implemented the same wait/scope/
display/notification helpers. Consolidate them without changing behavior.

- Add src/commands/merge_shared.rs: one WaitResult, one wait_for_pr_ready
  (blocked-message style parameterized so each command keeps its exact
  text), one record_ci_history_for_branch, shared strip_ansi/display_width,
  shared print_header family, a shared Scope + calculate_scope, and the
  PR-retarget cluster (moved off the merge command so siblings depend on a
  neutral module).
- Move desktop notifications into src/notifications.rs next to sound
  playback, with a pure notification_script seam + escaping test.
- Add scope unit tests for merge_remote and merge_queue (previously zero
  tests on both paths).

merge_rebase.rs, cli dispatch, flags, and all user-facing strings unchanged.
A shared merge-loop engine was deliberately declined as behavior-risky.
