+++ src/commands/merge_when_ready.rs
@@ -1,16 +1,17 @@
-use crate::commands::ci::{fetch_ci_statuses, record_ci_history};
-use crate::commands::merge::{
-    PrBaseUpdate, print_native_stack_locked_note, rebase_and_finalize_remaining_branch,
-    sync_head_after_push, update_pr_base_unless_current,
-};
 use crate::commands::merge_rebase::{
     fetch_remote_for_descendant_rebase, rebase_descendant_onto_remote_trunk_with_provenance,
+};
+use crate::commands::merge_shared::{
+    BlockedReasonStyle, PrBaseUpdate, WaitResult, calculate_scope, print_header,
+    print_header_error, print_header_success, print_native_stack_locked_note,
+    rebase_and_finalize_remaining_branch, record_ci_history_for_branch, sync_head_after_push,
+    update_pr_base_unless_current, wait_for_pr_ready,
 };
 use crate::config::Config;
 use crate::engine::Stack;
 use crate::forge::ForgeClient;
 use crate::git::{GitRepo, RebaseResult};
-use crate::github::pr::{MergeMethod, PrMergeStatus};
+use crate::github::pr::MergeMethod;
 use crate::ops::receipt::{OpKind, PlanSummary};
 use crate::ops::tx::{self, Transaction};
 use crate::progress::LiveTimer;
@@ -18,9 +19,8 @@ use crate::remote::RemoteInfo;
 use anyhow::{Context, Result};
 use colored::Colorize;
 use dialoguer::{Confirm, theme::ColorfulTheme};
-use std::io::Write;
 use std::process::Command;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 /// Information about a branch in the land scope
 #[derive(Debug, Clone)]
@@ -81,13 +81,6 @@ impl LandStatus {
             LandStatus::Failed(reason) => format!("failed: {}", reason).red().to_string(),
         }
     }
-}
-
-/// Result of waiting for a PR to be ready
-enum WaitResult {
-    Ready,
-    Failed(String),
-    Timeout,
 }
 
 #[allow(clippy::too_many_arguments)]
@@ -284,8 +277,16 @@ pub fn run(
                 print_dashboard(&branches, quiet);
             }
 
-            match wait_for_pr_ready(&rt, &client, pr_number, timeout, poll_interval, quiet)? {
-                WaitResult::Ready => {}
+            match wait_for_pr_ready(
+                &rt,
+                &client,
+                pr_number,
+                timeout,
+                poll_interval,
+                BlockedReasonStyle::StatusText,
+                quiet,
+            )? {
+                WaitResult::Ready(_) => {}
                 WaitResult::Failed(reason) => {
                     branches[idx].status = LandStatus::Failed(reason.clone());
                     failed_pr = Some((branch_name, pr_number, reason));
@@ -587,7 +588,7 @@ pub fn run(
         }
 
         // Send macOS notification
-        send_notification(
+        crate::notifications::send_desktop_notification(
             "stax merge --when-ready",
             &format!(
                 "Merged {} {} into {}",
@@ -647,28 +648,13 @@ fn calculate_merge_scope(
     all: bool,
     downstack_only: bool,
 ) -> MergeWhenReadyScope {
-    let mut to_merge = stack.ancestors(current);
-    to_merge.reverse();
-    to_merge.retain(|b| b != &stack.trunk);
-
-    let mut remaining = stack.descendants(current);
-    if downstack_only {
-        remaining.insert(0, current.to_string());
-    } else {
-        to_merge.push(current.to_string());
-    }
-
-    if all && !remaining.is_empty() {
-        to_merge.extend(remaining);
-        remaining = Vec::new();
-    }
-
+    let scope = calculate_scope(stack, current, all, downstack_only);
     MergeWhenReadyScope {
-        to_merge,
-        remaining,
-        trunk: stack.trunk.clone(),
-        current: current.to_string(),
-        downstack_only,
+        to_merge: scope.to_merge,
+        remaining: scope.remaining,
+        trunk: scope.trunk,
+        current: scope.current,
+        downstack_only: scope.downstack_only,
     }
 }
 
@@ -737,177 +723,6 @@ fn print_dashboard(branches: &[LandBranchInfo], quiet: bool) {
             println!("      {} {}", branch.status.symbol(), status_str.dimmed());
         }
     }
-}
-
-/// Wait for a PR to be ready to merge (CI passed, approved)
-fn wait_for_pr_ready(
-    rt: &tokio::runtime::Runtime,
-    client: &ForgeClient,
-    pr_number: u64,
-    timeout: Duration,
-    poll_interval: Duration,
-    quiet: bool,
-) -> Result<WaitResult> {
-    let start = Instant::now();
-    let mut last_status: Option<String> = None;
-
-    loop {
-        let status: PrMergeStatus =
-            rt.block_on(async { client.get_pr_merge_status(pr_number).await })?;
-
-        if status.is_ready() {
-            if !quiet && last_status.is_some() {
-                println!();
-            }
-            return Ok(WaitResult::Ready);
-        }
-
-        if status.is_blocked() {
-            if !quiet && last_status.is_some() {
-                println!();
-            }
-            return Ok(WaitResult::Failed(status.status_text().to_string()));
-        }
-
-        if start.elapsed() > timeout {
-            if !quiet && last_status.is_some() {
-                println!();
-            }
-            return Ok(WaitResult::Timeout);
-        }
-
-        // Show waiting status
-        if !quiet {
-            let elapsed = start.elapsed().as_secs();
-            let status_text = format!(
-                "      {} Waiting for {}... ({}s)",
-                "⏳".yellow(),
-                status.status_text().to_lowercase(),
-                elapsed
-            );
-
-            if last_status.is_some() {
-                print!("\r{}\r", " ".repeat(80));
-            }
-            print!("{}", status_text);
-            std::io::stdout().flush().ok();
-            last_status = Some(status_text);
-        }
-
-        std::thread::sleep(poll_interval);
-    }
-}
-
-/// Record CI history for a single branch after it's merged
-fn record_ci_history_for_branch(
-    repo: &GitRepo,
-    rt: &tokio::runtime::Runtime,
-    client: &ForgeClient,
-    stack: &Stack,
-    branch: &str,
-) {
-    if repo.branch_commit(branch).is_err() {
-        return;
-    }
-
-    let branches = vec![branch.to_string()];
-    if let Ok(statuses) = fetch_ci_statuses(repo, rt, client, stack, &branches) {
-        record_ci_history(repo, &statuses);
-    }
-}
-
-/// Send a macOS desktop notification
-fn send_notification(title: &str, message: &str) {
-    if cfg!(target_os = "macos") {
-        let script = format!(
-            r#"display notification "{}" with title "{}""#,
-            message.replace('"', "\\\""),
-            title.replace('"', "\\\""),
-        );
-        let _ = Command::new("osascript").args(["-e", &script]).output();
-    }
-}
-
-/// Strip ANSI codes for length calculation
-fn strip_ansi(s: &str) -> String {
-    let mut result = String::with_capacity(s.len());
-    let mut in_escape = false;
-
-    for c in s.chars() {
-        if c == '\x1b' {
-            in_escape = true;
-            continue;
-        }
-        if in_escape {
-            if c == 'm' {
-                in_escape = false;
-            }
-            continue;
-        }
-        result.push(c);
-    }
-
-    result
-}
-
-fn display_width(s: &str) -> usize {
-    let stripped = strip_ansi(s);
-    stripped
-        .chars()
-        .map(|c| match c {
-            '\x00'..='\x1f' | '\x7f' => 0,
-            '\x20'..='\x7e' => 1,
-            '─' | '│' | '┌' | '┐' | '└' | '┘' | '├' | '┤' | '┬' | '┴' | '┼' | '╭' | '╮' | '╯'
-            | '╰' | '║' | '═' => 1,
-            '←' | '→' | '↑' | '↓' => 1,
-            '✓' | '✗' | '✔' | '✘' => 1,
-            _ => 2,
-        })
-        .sum()
-}
-
-fn print_header(title: &str) {
-    let width: usize = 56;
-    let title_width = display_width(title);
-    let padding = width.saturating_sub(title_width) / 2;
-    println!("╭{}╮", "─".repeat(width));
-    println!(
-        "│{}{}{}│",
-        " ".repeat(padding),
-        title.bold(),
-        " ".repeat(width.saturating_sub(padding + title_width))
-    );
-    println!("╰{}╯", "─".repeat(width));
-}
-
-fn print_header_success(title: &str) {
-    let width: usize = 56;
-    let full_title = format!("✓ {}", title);
-    let title_width = display_width(&full_title);
-    let padding = width.saturating_sub(title_width) / 2;
-    println!("╭{}╮", "─".repeat(width));
-    println!(
-        "│{}{}{}│",
-        " ".repeat(padding),
-        full_title.green().bold(),
-        " ".repeat(width.saturating_sub(padding + title_width))
-    );
-    println!("╰{}╯", "─".repeat(width));
-}
-
-fn print_header_error(title: &str) {
-    let width: usize = 56;
-    let full_title = format!("✗ {}", title);
-    let title_width = display_width(&full_title);
-    let padding = width.saturating_sub(title_width) / 2;
-    println!("╭{}╮", "─".repeat(width));
-    println!(
-        "│{}{}{}│",
-        " ".repeat(padding),
-        full_title.red().bold(),
-        " ".repeat(width.saturating_sub(padding + title_width))
-    );
-    println!("╰{}╯", "─".repeat(width));
 }
 
 #[cfg(test)]
@@ -1026,20 +841,6 @@ mod tests {
         assert_eq!(info.branch, "feature-test");
         assert_eq!(info.pr_number, 42);
         assert_eq!(info.status, LandStatus::Pending);
-    }
-
-    #[test]
-    fn test_strip_ansi() {
-        assert_eq!(strip_ansi(""), "");
-        assert_eq!(strip_ansi("hello"), "hello");
-        assert_eq!(strip_ansi("\x1b[31mred\x1b[0m"), "red");
-    }
-
-    #[test]
-    fn test_display_width() {
-        assert_eq!(display_width("hello"), 5);
-        assert_eq!(display_width("✓"), 1);
-        assert_eq!(display_width("\x1b[32m✓\x1b[0m passed"), 8);
     }
 
     #[test]
