+++ src/commands/mod.rs
@@ -29,6 +29,7 @@ pub mod merge;
 pub mod merge_queue;
 pub(crate) mod merge_rebase;
 pub mod merge_remote;
+pub(crate) mod merge_shared;
 pub mod merge_stack;
 pub mod merge_when_ready;
 pub mod modify;
